refactor: move shared to another package
