+++ packages/graphql/src/graphql/workspace-revoke-permission.gql
@@ -1,0 +1,3 @@
+mutation WorkspaceRevokePermission($workspaceId: String!) {
+  p
+}
