+++ packages/workspace/package.json
@@ -7,6 +7,7 @@
     "./type": "./src/type.ts",
     "./migration": "./src/migration/index.ts",
     "./local/crud": "./src/local/crud.ts",
+    "./affine/gql": "./src/affine/gql.ts",
     "./providers": "./src/providers/index.ts"
   },
   "peerDependencies": {
@@ -17,6 +18,7 @@
     "@affine-test/fixtures": "workspace:*",
     "@affine/debug": "workspace:*",
     "@affine/env": "workspace:*",
+    "@affine/graphql": "workspace:*",
     "@toeverything/hooks": "workspace:*",
     "@toeverything/plugin-infra": "workspace:*",
     "@toeverything/y-indexeddb": "workspace:*",
